#34 rename: 폴더 구조 변경에 따른 import 이름 변경
