+++ src/main/java/com/gdsc/pikpet/repository/AnimalRepository.java
@@ -1,6 +1,6 @@
 package com.gdsc.pikpet.repository;
 
-import com.gdsc.pikpet.entity.Animal;
+import com.gdsc.pikpet.entity.animal.Animal;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.rest.core.annotation.RepositoryRestResource;
